kubeflow-preprocessor/pyproject.toml
[tool.poetry]
name = "kubeflow-preprocessor"
version = "0.1.0"
description = ""
authors = ["ya-mori <ppap.yohei.moriya@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.2.4"
sklearn = "^0.0"
tensorflow-cpu = "^2.5.0"
numpy = "1.19.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
